+++ scripts/generate_api.py
@@ -8,6 +8,10 @@ import subprocess
 import sys
 
 
+def normalize_version(version: str) -> str:
+    return re.sub(r"\.([a-zA-Z]+)(\d+)$", r"-\1.\2", version)
+
+
 if __name__ == "__main__":
     # Deleting existing generated stuff to avoid conflict
     print("Deleting directory...")
@@ -20,9 +24,9 @@ if __name__ == "__main__":
     print("Creating schema...")
     subprocess.call([sys.executable, os.path.join(__file__, "..", "export_openapi.py")])
 
-    # Extract the version and transform it from, for example "0.1.5dev2" (python notation) to "0.1.5-dev.2" (node notation)
+    # Extract the version and transform it from, for example "0.1.5.dev2" (python notation) to "0.1.5-dev.2" (node notation)
     with open("openapi.json", "r") as f:
-        _version = json.loads(f.read())["info"]["version"]
+        _version = normalize_version(json.loads(f.read())["info"]["version"])
 
     # Generate using openapi client
     print(f"Generating package version {_version}")