+++ src/rlgym_learn_api/desc/project/project.py
@@ -5,3 +5,4 @@ class ProjectMetadata(BaseModel):
     name: str
     id: str
     interpreter: str
+    created_at_version: str